Extracted pagehandler as extension method
