+++ Telerik.Sitefinity.Frontend/Mvc/Helpers/ResourceHelper.cs
@@ -267,7 +267,7 @@ namespace Telerik.Sitefinity.Frontend.Mvc.Helpers
             }
             else
             {
-                var page = PageInitializer.GetPageHandler(HttpContext.Current.Handler) ?? new PageProxy(null);
+                var page = HttpContext.Current.Handler.GetPageHandler() ?? new PageProxy(null);
 
                 resourceUrl = page.ClientScript.GetWebResourceUrl(
                     TypeResolutionService.ResolveType("Telerik.Sitefinity.Resources.Reference"),
@@ -320,7 +320,7 @@ namespace Telerik.Sitefinity.Frontend.Mvc.Helpers
             MvcHtmlString result = MvcHtmlString.Empty;
 
             // No section name renders the script inline if it hasn't been rendered
-            if (string.IsNullOrEmpty(sectionName) || !SectionRenderer.IsAvailable(PageInitializer.GetPageHandler(httpContext.Handler), sectionName))
+            if (string.IsNullOrEmpty(sectionName) || !SectionRenderer.IsAvailable(httpContext.Handler.GetPageHandler(), sectionName))
             {
                 if (!register.IsRegistered(resourcePath, sectionName: null))
                 {
@@ -383,7 +383,7 @@ namespace Telerik.Sitefinity.Frontend.Mvc.Helpers
 
         private static Page GetTopMostPageFromHandler(IHttpHandler httpHandler)
         {
-            var page = PageInitializer.GetPageHandler(httpHandler);
+            var page = httpHandler.GetPageHandler();
 
             if (page != null)
             {
